Create hangman.py
The program will generate a random word from the wordlist for the user to guess. 

Incorrectly guessing a letter will lose a guess, incorrectly guessing a vowel will lose 2 guesses. Imputing a '*' while provide a list of words that might be the word the user is guessing, but it is recommended that the user is close to figuring out the word (having most of the letters guessed correctly) before using the '*' element. Doing so when there are fewer correctly guessed words can make the list excessively long.